+++ Bricouli/src/main/java/com/demo/Bricouli/repositories/DevisRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.demo.Bricouli.models.Devis;
 
-public interface DevisRepository extends JpaRepository<Devis, Double>{
+public interface DevisRepository extends JpaRepository<Devis, Long>{
 
 }
